# Refactor cli tests