--- a/spec/support/gems/baz/lib/baz/role.rb
+++ b/spec/support/gems/baz/lib/baz/role.rb
@@ -1,9 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module Baz
-  class Role
-    include SmartProperties
-    property :title, accepts: String
-  end
-end