--- a/spec/support/gems/foo/lib/foo/secret.rb
+++ b/spec/support/gems/foo/lib/foo/secret.rb
@@ -1,8 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-module Foo
-  class Secret
-    VALUE = 42
-  end
-end