--- a/spec/support/repo/.gitignore
+++ b/spec/support/repo/.gitignore
@@ -1,3 +1,0 @@
-.bundle/
-Gemfile.lock
-sorbet/