--- a/spec/support/repo/config/environment.rb
+++ b/spec/support/repo/config/environment.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require_relative "application.rb"