--- a/spec/support/repo/lib/comment.rb
+++ b/spec/support/repo/lib/comment.rb
@@ -1,9 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-module Namespace
-  class Comment
-    include SmartProperties
-    property! :body, accepts: String
-  end
-end