--- a/spec/support/repo/lib/post.rb
+++ b/spec/support/repo/lib/post.rb
@@ -1,7 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-class Post
-  include SmartProperties
-  property :title, accepts: String
-end