--- a/spec/support/repo/lib/user.rb
+++ b/spec/support/repo/lib/user.rb
@@ -1,5 +1,0 @@
-# typed: true
-# frozen_string_literal: true
-
-class User
-end