--- a/spec/support/repo/postrequire.rb
+++ b/spec/support/repo/postrequire.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require "foo/secret"