--- a/spec/support/repo/postrequire_faulty.rb
+++ b/spec/support/repo/postrequire_faulty.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-
-require "foo/will_fail"