--- a/spec/support/require/multi/file3.rb
+++ b/spec/support/require/multi/file3.rb
@@ -1,3 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-require "e"