--- a/spec/support/require/multi/lib/file1.rb
+++ b/spec/support/require/multi/lib/file1.rb
@@ -1,3 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-require "a"