--- a/spec/support/require/multi/sorbet/config
+++ b/spec/support/require/multi/sorbet/config
@@ -1,3 +1,0 @@
-lib/
-file1.rb
-file2.rb