--- a/spec/support/require/project_ignore/file1.rb
+++ b/spec/support/require/project_ignore/file1.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-require "a"
-require "libc"