--- a/spec/support/require/project_ignore/file2.rb
+++ b/spec/support/require/project_ignore/file2.rb
@@ -1,4 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-require "b"
-require "libd"