--- a/spec/support/require/project_ignore/lib/a.rb
+++ b/spec/support/require/project_ignore/lib/a.rb
@@ -1,3 +1,0 @@
-# typed: strict
-# frozen_string_literal: true
-require "liba"