--- a/spec/support/require/simple/sorbet/config
+++ b/spec/support/require/simple/sorbet/config
@@ -1,1 +1,0 @@
-simple.rb