--- a/spec/support/require/sorbet_ignore/sorbet/config
+++ b/spec/support/require/sorbet_ignore/sorbet/config
@@ -1,3 +1,0 @@
-.
---ignore=lib/
---ignore=file3.rb