--- a/spec/tapioca/cli/deprecations_spec.rb
+++ b/spec/tapioca/cli/deprecations_spec.rb
@@ -1,24 +1,28 @@
-# typed: true
+# typed: strict
 # frozen_string_literal: true
 
-require "cli_spec"
+require "spec_with_project"
 
 module Tapioca
-  class DeprecationsSpec < CliSpec
-    describe("deprecations") do
+  class DeprecationsSpec < SpecWithProject
+    describe("#cli::dsl::deprecations") do
+      before(:all) do
+        project.bundle_install
+      end
+
       it "prints the correct deprecation message with -c" do
-        output = tapioca("dsl -c foo")
-        assert_includes(output, "DEPRECATION: The `-c` and `--cmd` flags will be removed in a future release.")
+        out, _, _ = @project.tapioca("dsl -c foo")
+        assert_includes(out, "DEPRECATION: The `-c` and `--cmd` flags will be removed in a future release.")
       end
 
       it "prints the correct deprecation message with --cmd" do
-        output = tapioca("dsl --cmd foo")
-        assert_includes(output, "DEPRECATION: The `-c` and `--cmd` flags will be removed in a future release.")
+        out, _, _ = @project.tapioca("dsl --cmd foo")
+        assert_includes(out, "DEPRECATION: The `-c` and `--cmd` flags will be removed in a future release.")
       end
 
       it "doesn't print the correct deprecation message with no flag" do
-        output = tapioca("dsl")
-        refute_includes(output, "DEPRECATION: The `-c` and `--cmd` flags will be removed in a future release.")
+        out, _, _ = @project.tapioca("dsl")
+        refute_includes(out, "DEPRECATION: The `-c` and `--cmd` flags will be removed in a future release.")
       end
     end
   end