--- a/spec/tapioca/cli/version_spec.rb
+++ b/spec/tapioca/cli/version_spec.rb
@@ -1,19 +1,27 @@
-# typed: true
+# typed: strict
 # frozen_string_literal: true
 
-require "cli_spec"
+require "spec_with_project"
 
 module Tapioca
-  class VersionSpec < CliSpec
-    describe("#version") do
+  class VersionSpec < SpecWithProject
+    describe("#cli::version") do
+      before(:all) do
+        project.bundle_install
+      end
+
       it "must display the version when passing --version" do
-        output = tapioca("--version")
-        assert_equal("Tapioca v#{Tapioca::VERSION}", output.strip)
+        out, err, status = @project.tapioca("--version")
+        assert_equal("Tapioca v#{Tapioca::VERSION}", out.strip)
+        assert_empty(err)
+        assert(status)
       end
 
       it "must display the version when passing -v" do
-        output = tapioca("-v")
-        assert_equal("Tapioca v#{Tapioca::VERSION}", output.strip)
+        out, err, status = @project.tapioca("-v")
+        assert_equal("Tapioca v#{Tapioca::VERSION}", out.strip)
+        assert_empty(err)
+        assert(status)
       end
     end
   end