… spec/support/gems/bar/bar.gemspec
… spec/support/gems/bar/lib/bar.rb
… spec/support/gems/baz/baz.gemspec
… spec/support/gems/baz/lib/baz.rb
… spec/support/gems/foo/foo.gemspec
… spec/support/gems/foo/lib/foo.rb
… spec/support/gems/qux/lib/qux.rb
… spec/support/gems/qux/qux.gemspec
… spec/support/repo/Gemfile
… spec/support/repo/bin/tapioca
… spec/support/repo/config/application.rb
… spec/support/repo/config/shim.rbi
… spec/support/repo/lib/database.rb
… spec/support/repo/lib/generators/foo/generator.rb
… spec/support/repo/lib/generators/generator_that_includes_bar_module_in_post.rb
… spec/support/repo/lib/generators/generator_that_includes_foo_module_in_post.rb
… spec/support/repo/lib/job.rb
… spec/support/require/empty/sorbet/config
… spec/support/require/project_ignore/sorbet/config
… spec/support/require/simple/simple.rb
… spec/tapioca/cli/dsl_spec.rb
… spec/tapioca/cli/gem_spec.rb
… spec/tapioca/cli/todo_spec.rb
… spec/tapioca/compilers/requires_compiler_spec.rb